M16CRYPTO-247 M16CRYPTO-248 M16CRYPTO-249 M16CRYPTO-250 MISRA for all SHA projects

+++ dspic33ak512mps512/sha2/dspic33ak512mps512-sha2.X/app/src/app_sha2.c
@@ -45,7 +45,7 @@ static void SHA2_Digest(const TEST_VECTOR* vector)
 
     headerOutputPrint(vector->algorithm, vector->vectorInformation);
 
-    if(vector->algorithm == SHA2_224_ALGO)
+    if(vector->algorithm == (int)SHA2_224_ALGO)
     {
         benchmarkingStart();
 
@@ -63,7 +63,7 @@ static void SHA2_Digest(const TEST_VECTOR* vector)
 
         benchmarkingEnd(&ticksToProcess);
     }
-    else if(vector->algorithm == SHA2_256_ALGO)
+    else if(vector->algorithm == (int)SHA2_256_ALGO)
     {
         benchmarkingStart();
 
@@ -81,7 +81,7 @@ static void SHA2_Digest(const TEST_VECTOR* vector)
 
         benchmarkingEnd(&ticksToProcess);
     }
-    else if(vector->algorithm == SHA2_384_ALGO)
+    else if(vector->algorithm == (int)SHA2_384_ALGO)
     {
         benchmarkingStart();
 
@@ -99,7 +99,7 @@ static void SHA2_Digest(const TEST_VECTOR* vector)
 
         benchmarkingEnd(&ticksToProcess);
     }
-    else if(vector->algorithm == SHA2_512_ALGO)
+    else if(vector->algorithm == (int)SHA2_512_ALGO)
     {
         benchmarkingStart();
 
@@ -122,8 +122,8 @@ static void SHA2_Digest(const TEST_VECTOR* vector)
         (void) printf(RED"\r\n Unsupported SHA2 Test Vector");
     }
 
-    if((vector->algorithm == SHA2_224_ALGO) || (vector->algorithm == SHA2_256_ALGO)
-            || (vector->algorithm == SHA2_384_ALGO) || (vector->algorithm == SHA2_512_ALGO))
+    if((vector->algorithm == (int)SHA2_224_ALGO) || (vector->algorithm == (int)SHA2_256_ALGO)
+            || (vector->algorithm == (int)SHA2_384_ALGO) || (vector->algorithm == (int)SHA2_512_ALGO))
     {
         printHexArray("Expected:      ", vector->digest, vector->digestSize);
         printHexArray("Result:        ", resultDigest, vector->digestSize);
